all JWT authenitcation and authorization are done

+++ controllers/contactController.js
@@ -3,7 +3,7 @@ const Contact = require("../models/contactModel");
 
 
 const getContact = asyncHandler(async(req, res) => {
-    const contact=await Contact.find()
+    const contact=await Contact.find({user_id: req.user.id})
     res.status(200).json(contact );
 });
 
@@ -14,7 +14,7 @@ const addContact = asyncHandler(async (req, res) => {
         return res.status(400).json({ message: "Name and email are required" });
     }
 
-   const contact=await Contact.create({name, email})
+   const contact=await Contact.create({name, email, user_id:req.user.id})
    res.status(200).json(contact)
 
 });
@@ -36,13 +36,19 @@ const updateContact = asyncHandler(async (req, res) => {
         res.status(404)
         throw new Error("contanct not found")
     }
+    if(contact.user_id !==req.user.id){
+        res.status(404)
+        throw new Error("unauthorized access to modify the data")
+    }
+ 
+
   const updatedContact=await Contact.findByIdAndUpdate(req.params.id,
     req.body,{new:true , runValidators:true}
   )
     res.status(200).json(updatedContact);
 });
 
-
+   
 const deleteContact = asyncHandler(async (req, res) => {
 const contact= await Contact.findById(req.params.id)
 if(!contact)
@@ -50,6 +56,11 @@ if(!contact)
     res.status(404)
     throw new Error("contact not found")
 }
+if(contact.user_id !==req.user.id){
+    res.status(404)
+    throw new Error("unauthorized access to modify the data")
+}
+
    await Contact.findByIdAndDelete(req.params.id)
    res.status(200).json(contact)
 });
